--- a/backend/app/Http/Controllers/Api/OrderController.php
+++ b/backend/app/Http/Controllers/Api/OrderController.php
@@ -3,14 +3,14 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
+use App\Models\CartItem;
 use App\Models\Order;
 use App\Models\OrderItem;
 use App\Models\Product;
-use App\Models\CartItem;
-use App\Services\NotificationService;
 use App\Services\InventoryService;
-use Illuminate\Http\Request;
+use App\Services\NotificationService;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Validation\ValidationException;
 
@@ -19,8 +19,7 @@ class OrderController extends Controller
     public function __construct(
         private NotificationService $notificationService,
         private InventoryService $inventoryService
-    ) {
-    }
+    ) {}
 
     /**
      * Display user's orders.
@@ -31,7 +30,7 @@ class OrderController extends Controller
             ->with(['orderItems.product.categories', 'orderItems.product.images', 'orderItems.product.producer'])
             ->orderBy('created_at', 'desc')
             ->get();
-            
+
         return response()->json([
             'orders' => $orders->map(function ($order) {
                 return [
@@ -65,7 +64,7 @@ class OrderController extends Controller
                         ];
                     }),
                 ];
-            })
+            }),
         ]);
     }
 
@@ -101,10 +100,11 @@ class OrderController extends Controller
 
             foreach ($cartItems as $cartItem) {
                 $product = $cartItem->product;
-                
+
                 // Check if product is still active
-                if (!$product->is_active) {
+                if (! $product->is_active) {
                     $unavailableProducts[] = $product->name;
+
                     continue;
                 }
 
@@ -113,8 +113,9 @@ class OrderController extends Controller
                     $insufficientStock[] = [
                         'product' => $product->name,
                         'requested' => $cartItem->quantity,
-                        'available' => $product->stock
+                        'available' => $product->stock,
                     ];
+
                     continue;
                 }
 
@@ -133,19 +134,19 @@ class OrderController extends Controller
             }
 
             // Handle validation errors
-            if (!empty($unavailableProducts)) {
+            if (! empty($unavailableProducts)) {
                 throw ValidationException::withMessages([
-                    'products' => ['The following products are no longer available: ' . implode(', ', $unavailableProducts)],
+                    'products' => ['The following products are no longer available: '.implode(', ', $unavailableProducts)],
                 ]);
             }
 
-            if (!empty($insufficientStock)) {
+            if (! empty($insufficientStock)) {
                 $stockErrors = [];
                 foreach ($insufficientStock as $stockError) {
                     $stockErrors[] = "{$stockError['product']}: requested {$stockError['requested']}, only {$stockError['available']} available";
                 }
                 throw ValidationException::withMessages([
-                    'stock' => ['Insufficient stock for: ' . implode('; ', $stockErrors)],
+                    'stock' => ['Insufficient stock for: '.implode('; ', $stockErrors)],
                 ]);
             }
 
@@ -233,7 +234,7 @@ class OrderController extends Controller
                             ] : null,
                         ];
                     }),
-                ]
+                ],
             ], 201);
 
         } catch (ValidationException $e) {
@@ -241,9 +242,10 @@ class OrderController extends Controller
             throw $e;
         } catch (\Exception $e) {
             DB::rollBack();
+
             return response()->json([
                 'message' => 'Failed to create order',
-                'error' => $e->getMessage()
+                'error' => $e->getMessage(),
             ], 500);
         }
     }
@@ -254,7 +256,7 @@ class OrderController extends Controller
     public function store(Request $request): JsonResponse
     {
         $this->authorize('create', Order::class);
-        
+
         $validated = $request->validate([
             'items' => 'required|array|min:1',
             'items.*.product_id' => 'required|exists:products,id',
@@ -277,7 +279,7 @@ class OrderController extends Controller
                     ->lockForUpdate() // Prevent race conditions on stock
                     ->first();
 
-                if (!$product) {
+                if (! $product) {
                     throw ValidationException::withMessages([
                         'items' => ["Product with ID {$item['product_id']} not found or inactive."],
                     ]);
@@ -393,9 +395,10 @@ class OrderController extends Controller
             throw $e;
         } catch (\Exception $e) {
             DB::rollBack();
+
             return response()->json([
                 'message' => 'Failed to create order',
-                'error' => $e->getMessage()
+                'error' => $e->getMessage(),
             ], 500);
         }
     }