--- a/backend/app/Http/Resources/ProducerResource.php
+++ b/backend/app/Http/Resources/ProducerResource.php
@@ -10,7 +10,6 @@ class ProducerResource extends JsonResource
     /**
      * Transform the resource into an array.
      *
-     * @param Request $request
      * @return array<string, mixed>
      */
     public function toArray(Request $request): array