… backend/app/Http/Controllers/Api/Producer/ProducerAnalyticsController.php
… backend/app/Http/Controllers/Api/ProducerController.php
… backend/app/Http/Controllers/Api/ProductController.php
… backend/app/Http/Controllers/Api/RefundController.php
… backend/app/Http/Controllers/Api/ShippingController.php
… backend/app/Http/Controllers/Api/V1/OrderController.php
… backend/app/Http/Controllers/Api/V1/ProducerController.php
… backend/app/Http/Controllers/Api/V1/ProductController.php
… backend/app/Http/Controllers/Api/WebhookController.php
… backend/app/Http/Controllers/Public/ProducerController.php
… backend/app/Http/Controllers/Public/ProductController.php
… backend/app/Http/Requests/StoreOrderRequest.php
… backend/app/Http/Requests/UpdateProducerRequest.php
… backend/app/Http/Requests/UpdateProductRequest.php
… backend/app/Http/Resources/OrderItemResource.php
… backend/app/Http/Resources/OrderResource.php
… backend/app/Http/Resources/ProductResource.php
… backend/app/Http/Resources/PublicProducerResource.php
… backend/app/Models/Message.php
… backend/app/Models/Notification.php
… backend/app/Models/Order.php
… backend/app/Models/Producer.php
… backend/app/Models/Product.php
… backend/app/Models/Shipment.php
… backend/app/Models/User.php
… backend/app/Policies/OrderPolicy.php
… backend/app/Policies/ProducerPolicy.php
… backend/app/Policies/ProductPolicy.php
… backend/app/Providers/AuthServiceProvider.php
… backend/app/Services/AnalyticsService.php